divide SnakeEncoding
