+++ Source/TonSwift/Util/String+CellEncoding.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension String {
+    /// Encodes a String into a Cell
+    public func toTonCell() throws -> Cell {
+        try Builder().writeSnakeData(Data(utf8)).endCell()
+    }
+}
+
